+++ app/src/main/java/woowacourse/movie/confirm/ReservationConfirmActivity.kt
@@ -1,9 +1,13 @@
-package woowacourse.movie
+package woowacourse.movie.confirm
 
 import android.os.Bundle
 import android.util.Log
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
+import woowacourse.movie.Movie
+import woowacourse.movie.R
+import woowacourse.movie.main.MainActivity
+import woowacourse.movie.reservation.MovieDetailActivity
 
 class ReservationConfirmActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -19,7 +23,7 @@ class ReservationConfirmActivity : AppCompatActivity() {
         val moneyTextView = findViewById<TextView>(R.id.reservation_money)
 
         titleTextView.text = movie.title
-        dateTextView.text = movie.date.toString()
+        dateTextView.text = movie.startDate.toString()
         moneyTextView.text = (reservationCount * MOVIE_TICKET_MONEY).toString()
     }
