+++ app/src/main/java/woowacourse/movie/domain/DiscountCalculator.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.domain
 
 import java.time.LocalDateTime
