+++ app/src/main/java/woowacourse/movie/main/MainActivity.kt
@@ -1,9 +1,12 @@
-package woowacourse.movie
+package woowacourse.movie.main
 
 import android.content.Intent
 import android.os.Bundle
 import android.widget.ListView
 import androidx.appcompat.app.AppCompatActivity
+import woowacourse.movie.Movie
+import woowacourse.movie.R
+import woowacourse.movie.reservation.MovieDetailActivity
 import java.time.LocalDate
 
 class MainActivity : AppCompatActivity() {
@@ -31,6 +34,7 @@ class MainActivity : AppCompatActivity() {
                     R.drawable.slamdunk,
                     "더 퍼스트 슬램덩크",
                     LocalDate.of(2023, 1, 4),
+                    LocalDate.of(2023, 2, 23),
                     124,
                     "북산고 농구부는 전국 대회에 출전해 라이벌 산왕공고와 맞붙는다. 멤버 각자가 쌓아온 성과, 그들이 짊어진 과거, 다양한 생각들이 뜨거운 코트 위에서 다시 한번 격렬하게 충돌한다." +
                         "북산고 농구부는 전국 대회에 출전해 라이벌 산왕공고와 맞붙는다. 멤버 각자가 쌓아온 성과, 그들이 짊어진 과거, 다양한 생각들이 뜨거운 코트 위에서 다시 한번 격렬하게 충돌한다." +
