+++ app/src/main/res/layout/activity_movie_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MovieDetailActivity">
+    tools:context=".reservation.MovieDetailActivity">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/begin_guide"
