+++ app/src/main/res/layout/activity_reservation_confirm.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ReservationConfirmActivity">
+    tools:context=".confirm.ReservationConfirmActivity">
 
     <TextView
         android:id="@+id/empty_top"
