+++ app/src/test/java/woowacourse/movie/RunningTimeSetterTest.kt
@@ -2,6 +2,7 @@ package woowacourse.movie
 
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
+import woowacourse.movie.domain.RunningTimeSetter
 import java.time.LocalDate
 import java.time.LocalTime
